+++ tests/runtime_change/test_runtime_change.nim
@@ -2,9 +2,11 @@ import lazy_rest, strutils
 
 const
   special_config = slurp("special.cfg")
+  nearly_empty_config = slurp("nearly_empty.cfg")
   expected_title = "Hello Rst!"
   expected_string = "Crayon pop rules"
   filename = "hello.rst"
+
 
 proc count_instances(text, substring: string): int =
   ## Returns the number of times `substring` appears in `text`.
@@ -12,6 +14,7 @@ proc count_instances(text, substring: string): int =
   while pos >= 0:
     result += 1
     pos = text.find(substring, start = pos + substring.len)
+
 
 proc normal_test() =
   # Expects the normal thing to happen.
@@ -21,14 +24,29 @@ proc normal_test() =
   doAssert normal_text.find(expected_string) < 0
   doAssert normal_count == 2
 
-proc test() =
-  normal_test()
+
+proc special_test() =
+  # Generates the file with a different configuration.
   let
     special_text = rst_file_to_html(filename, special_config.parse_rst_options)
     special_count = special_text.count_instances(expected_title)
   doAssert special_text.find(expected_string) > 0
   doAssert special_count == 4
+  write_file("special.html", special_text)
+
+
+proc empty_test() =
+  # Runs with a nearly empty file, to verify the embedded one works.
+  let text = filename.rst_file_to_html(nearly_empty_config.parse_rst_options)
+  write_file("nearly_empty.html", text)
+
+
+proc test() =
   normal_test()
+  special_test()
+  normal_test()
+  empty_test()
+
 
 when isMainModule:
   test()