+++ docs/changes.rst
@@ -25,3 +25,5 @@ v0.1.0, ????-??-??
   <https://github.com/gradha/lazy_rest/issues/11>`_.
 * `Added customisation of file include and message procs
   <https://github.com/gradha/lazy_rest/issues/17>`_.
+* `Modified config options to be additive
+  <https://github.com/gradha/lazy_rest/issues/21>`_.